+++ CHANGELOG.md
@@ -266,6 +266,7 @@ Only display line where quantity is greater than 0.
 - Tax report : Refactoring entirely the debit tax report (type 12) and display the move lines details.
 - Accounting reports : disable page break interval to get only one tab on excel files
 - EBICS partner : Add partner service for bank statement file format to manage specific bank codification.
+- FIXED ASSET BATCH : Move Fixed asset Batch to Accounting batch.
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.